fix lỗi 404 error

+++ database/migrations/2020_04_15_163111_create_movies_table.php
@@ -27,8 +27,8 @@ class CreateMoviesTable extends Migration
             $table->integer('country')->comment('quoc gia');
             $table->integer('rate')->comment('diem danh gia');
             $table->string('slug')->comment('ten');
-            $table->integer('user_created_id')->comment('nguoi tao');
-            $table->integer('user_updated_id')->comment('nguoi cap nhat');
+            $table->unsignedBigInteger('user_created_id')->comment('nguoi tao');
+            $table->unsignedBigInteger('user_updated_id')->comment('nguoi cap nhat');
             $table->timestamps();
             $table->softDeletes()->comment('thoi gian xoa');
         });
