+++ database/migrations/2020_04_15_163556_create_tags_table.php
@@ -17,8 +17,8 @@ class CreateTagsTable extends Migration
        $table->bigIncrements('id')->comment('id');
        $table->string('name')->comment('ten');
        $table->string('slug')->nullable()->comment('duong dan');
-       $table->integer('user_created_id')->comment('nguoi tao');
-       $table->integer('user_updated_id')->comment('nguoi cap nhat');
+       $table->unsignedBigInteger('user_created_id')->comment('nguoi tao');
+       $table->unsignedBigInteger('user_updated_id')->comment('nguoi cap nhat');
        $table->timestamps();
        $table->softDeletes()->comment('thoi gian xoa');
      });
