+++ database/migrations/2020_04_15_163807_create_types_table.php
@@ -19,8 +19,8 @@ class CreateTypesTable extends Migration
             $table->text('description')->comment('mo ta');
             $table->string('slug')->comment('duong dan');
             $table->string('table_name')->comment('ten cua bang co the loai');
-            $table->integer('user_created_id')->comment('nguoi tao');
-            $table->integer('user_updated_id')->comment('nguoi cap nhat');
+            $table->unsignedBigInteger('user_created_id')->comment('nguoi tao');
+            $table->unsignedBigInteger('user_updated_id')->comment('nguoi cap nhat');
             $table->timestamps();
             $table->softDeletes()->comment('thoi gian xoa');
         });
