+++ database/migrations/2020_04_15_164019_create_videos_table.php
@@ -18,13 +18,13 @@ class CreateVideosTable extends Migration
             $table->string('title')->comment('mo ta');
             $table->string('tags')->comment('the tag');
             $table->text('description')->comment('mo ta');
-            $table->integer('movie_id')->comment('id cua movie');
-            $table->integer('playlist_id')->nullable()->comment('id cua playlist');
+            $table->unsignedBigInteger('movie_id')->comment('id cua movie');
+            $table->unsignedBigInteger('playlist_id')->nullable()->comment('id cua playlist');
             $table->integer('status')->default(1)->comment('trang thai');
             $table->integer('chap')->nullable()->comment('tap so may');
             $table->string('slug')->comment('duong dan');
-            $table->integer('user_created_id')->comment('nguoi tao');
-            $table->integer('user_updated_id')->comment('nguoi cap nhat');
+            $table->unsignedBigInteger('user_created_id')->comment('nguoi tao');
+            $table->unsignedBigInteger('user_updated_id')->comment('nguoi cap nhat');
             $table->timestamps();
             $table->softDeletes()->comment('thoi gian xoa');
         });
