+++ database/migrations/2020_04_15_164212_create_options_values_table.php
@@ -15,13 +15,13 @@ class CreateOptionsValuesTable extends Migration
     {
         Schema::create('option_values', function (Blueprint $table) {
             $table->bigIncrements('id')->comment('id');
-            $table->integer('option_id')->comment('id cua option');
+            $table->unsignedBigInteger('option_id')->comment('id cua option');
             $table->string('name')->comment('ten');
             $table->integer('order')->comment('so thu tu');
             $table->text('description')->nullable()->comment('mo ta');
             $table->integer('status')->default(1)->comment('trang thai');
-            $table->integer('user_created_id')->comment('nguoi tao');
-            $table->integer('user_updated_id')->comment('nguoi cap nhat');
+            $table->unsignedBigInteger('user_created_id')->comment('nguoi tao');
+            $table->unsignedBigInteger('user_updated_id')->comment('nguoi cap nhat');
             $table->timestamps();
             $table->softDeletes()->comment('thoi gian xoa');
         });
