+++ database/migrations/2020_04_15_164415_create_roles_table.php
@@ -20,8 +20,8 @@ class CreateRolesTable extends Migration
             $table->text('description')->comment('mo ta');
 
             $table->integer('status')->default(1)->comment('trang thai');
-            $table->integer('user_created_id')->comment('nguoi tao');
-            $table->integer('user_updated_id')->comment('nguoi cap nhat');
+            $table->unsignedBigInteger('user_created_id')->comment('nguoi tao');
+            $table->unsignedBigInteger('user_updated_id')->comment('nguoi cap nhat');
             $table->timestamps();
             $table->softDeletes()->comment('thoi gian xoa');
         });
