+++ database/migrations/2020_04_28_035920_create_manga_ads_table.php
@@ -15,12 +15,12 @@ class CreateMangaAdsTable extends Migration
     {
         Schema::create('manga_ads', function (Blueprint $table) {
             $table->bigIncrements('id')->comment('id');
-            $table->integer('object_id')->comment('id cua object');
+            $table->unsignedBigInteger('object_id')->comment('id cua object');
             $table->string('table_name')->comment('ten bang so huu');
             $table->string('link')->comment('link quang cao');
             $table->string('artical')->comment('link anh quang cao');  
-            $table->integer('user_created_id')->comment('nguoi tao');
-            $table->integer('user_updated_id')->comment('nguoi cap nhat');
+            $table->unsignedBigInteger('user_created_id')->comment('nguoi tao');
+            $table->unsignedBigInteger('user_updated_id')->comment('nguoi cap nhat');
             $table->timestamps();
             $table->softDeletes()->comment('thoi gian xoa');
         });
